If the user toggles light/dark mode, regenerate rasterised map images.
Previously, when toggling this mode the CSS styles would be updated, but the images would remain in the previous colour mode. Therefore an incorrect map color scheme was visible when zoomed out. Now the images will be regenerated so the colour scheme matches again.
