+++ client/src/App.tsx
@@ -2,11 +2,14 @@ import { QueryClientProvider } from '@tanstack/react-query';
 import queryClient from './api/queryClient';
 import { GameContextProvider } from './components/context/GameContext';
 import AppRoot from './components/AppRoot';
+import { DarkModeContextContextProvider } from './components/context/DarkModeContext';
 
 const App = () => (
   <QueryClientProvider client={queryClient}>
     <GameContextProvider>
-      <AppRoot />
+      <DarkModeContextContextProvider>
+        <AppRoot />
+      </DarkModeContextContextProvider>
     </GameContextProvider>
   </QueryClientProvider>
 );
